chore(kubernetes): test working provider
