+++ prowler/providers/kubernetes/services/apiserver/apiserver_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.apiserver.apiserver_service import APIServer
 
-apiserver_client = APIServer(global_provider)
+apiserver_client = APIServer(get_global_provider())
