+++ prowler/providers/kubernetes/services/controllermanager/controllermanager_client.py
@@ -1,6 +1,6 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.controllermanager.controllermanager_service import (
     ControllerManager,
 )
 
-controllermanager_client = ControllerManager(global_provider)
+controllermanager_client = ControllerManager(get_global_provider())
