+++ prowler/providers/kubernetes/services/core/core_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.core.core_service import Core
 
-core_client = Core(global_provider)
+core_client = Core(get_global_provider())
