+++ prowler/providers/kubernetes/services/etcd/etcd_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.etcd.etcd_service import Etcd
 
-etcd_client = Etcd(global_provider)
+etcd_client = Etcd(get_global_provider())
