+++ prowler/providers/kubernetes/services/kubelet/kubelet_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.kubelet.kubelet_service import Kubelet
 
-kubelet_client = Kubelet(global_provider)
+kubelet_client = Kubelet(get_global_provider())
