+++ prowler/providers/kubernetes/services/kubelet/kubelet_service.py
@@ -1,14 +1,15 @@
 import yaml
 
 from prowler.lib.logger import logger
+from prowler.providers.kubernetes.kubernetes_provider import KubernetesProvider
 from prowler.providers.kubernetes.lib.service.service import KubernetesService
 from prowler.providers.kubernetes.services.core.core_client import core_client
 
 
 ################## Kubelet ##################
 class Kubelet(KubernetesService):
-    def __init__(self, audit_info):
-        super().__init__(audit_info)
+    def __init__(self, provider: KubernetesProvider):
+        super().__init__(provider)
         self.client = core_client
 
         self.kubelet_config_maps = self.__get_kubelet_config_maps__()
