+++ prowler/providers/kubernetes/services/rbac/rbac_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.rbac.rbac_service import Rbac
 
-rbac_client = Rbac(global_provider)
+rbac_client = Rbac(get_global_provider())
