+++ prowler/providers/kubernetes/services/scheduler/scheduler_client.py
@@ -1,4 +1,4 @@
-from prowler.providers.common.common import global_provider
+from prowler.providers.common.common import get_global_provider
 from prowler.providers.kubernetes.services.scheduler.scheduler_service import Scheduler
 
-scheduler_client = Scheduler(global_provider)
+scheduler_client = Scheduler(get_global_provider())
